# neus-ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenApi Specification ",
        "description": "OpenApi Documentation for Alem Equb",
        "termsOfService": "term of service",
        "contact": {
            "name": "Habtamu",
            "email": "habtamuadugna.d@gmail.com"
        },
        "license": {
            "name": "Licence name ",
            "url": "http://url licence"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api",
            "description": "Local ENV"
        }
    ],
    "security": [
        {
            "keycloak": []
        }
    ],
    "paths": {
        "/users/{user-id}": {
            "put": {
                "tags": [
                    "users"
                ],
                "operationId": "toggleUserStatus",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user-exam": {
            "put": {
                "tags": [
                    "user-exam"
                ],
                "operationId": "saveUserAnswers",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserExamDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/save/{email}": {
            "post": {
                "tags": [
                    "users"
                ],
                "operationId": "saveUser",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/page": {
            "get": {
                "tags": [
                    "users"
                ],
                "operationId": "getPageOfUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseUserDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user-exam/{exam-id}": {
            "get": {
                "tags": [
                    "user-exam"
                ],
                "operationId": "getUserAnswers",
                "parameters": [
                    {
                        "name": "exam-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserExamDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/exams/{exam-id}": {
            "get": {
                "tags": [
                    "exams"
                ],
                "operationId": "getExamById",
                "parameters": [
                    {
                        "name": "exam-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "exams"
                ],
                "operationId": "deleteExam",
                "parameters": [
                    {
                        "name": "exam-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CorrectAnswer": {
                "type": "object",
                "properties": {
                    "questionId": {
                        "type": "string"
                    },
                    "isCorrect": {
                        "type": "boolean"
                    }
                }
            },
            "UpdateUserExamDto": {
                "type": "object",
                "properties": {
                    "examId": {
                        "type": "string"
                    },
                    "mode": {
                        "type": "string",
                        "enum": [
                            "STUDY",
                            "TEST"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "IN_PROGRESS",
                            "COMPLETED"
                        ]
                    },
                    "timeLeftInMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "userAnswers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserAnswer"
                        }
                    },
                    "correctAnswers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CorrectAnswer"
                        }
                    }
                }
            },
            "UserAnswer": {
                "type": "object",
                "properties": {
                    "questionId": {
                        "type": "string"
                    },
                    "choiceId": {
                        "type": "string"
                    }
                }
            },
...
            "UserDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "subscriptionLevel": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "registeredDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "enabled": {
                        "type": "boolean"
                    }
                }
            },
            "PageResponseUserDto": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserDto"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "UserExamDto": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string"
                    },
                    "examId": {
                        "type": "string"
                    },
                    "mode": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "IN_PROGRESS",
                            "COMPLETED"
                        ]
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "timeLeftInMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "studyModeUserAnswers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserAnswer"
                        }
                    },
                    "studyModeCorrectAnswers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CorrectAnswer"
                        }
                    },
                    "testModeUserAnswers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserAnswer"
                        }
                    },
                    "testModeCorrectAnswers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CorrectAnswer"
                        }
                    }
                }
            },
...
            "ExamDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "examType": {
                        "type": "string",
                        "enum": [
                            "TEST",
                            "ERMP",
                            "USMLE_STEP_1",
                            "USMLE_STEP_2"
                        ]
                    },
                    "year": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "duration": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "numberOfQuestions": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "requiredSubLevel": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    },
                    "randomQuestionCount": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "keycloak": {
                "type": "oauth2",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:9090/realms/neus/protocol/openid-connect/auth",
                        "tokenUrl": "http://localhost:9090/realms/neus/protocol/openid-connect/token",
                        "scopes": {}
                    }
                }
            }
        }
    }
}